make micro_release does the same as make misc-clean and make -C src/ clean, plus a bit of other cleanup such as stripping binaries.  Also only perform build cleanup on release versions.  For the develop version leave everything just as though sage has just been built.

+++ sagemath/scripts/install_sage.sh
@@ -55,7 +55,11 @@ EOFSAGE
 #
 # Unfortunately none of the existing make targets for sage cover this ground
 # exactly
-make misc-clean
-make -C src/ clean
-rm -rf upstream/
-rm -rf src/doc/output/doctrees/
+
+# For the 'develop' image we leave everything as it would be after a
+# successful sage build
+if [ $BRANCH != "develop" ]; then
+  make micro_release
+  rm -rf upstream/
+  rm -rf src/doc/output/doctrees/
+fi
